backend/requirements.txt
# ===== FastAPI Core =====
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ===== Database =====
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ===== Redis / Caching =====
redis[asyncio]==5.0.4
hiredis==2.3.2

# ===== Auth =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ===== Validation =====
pydantic==2.7.1
pydantic-settings==2.2.1

# ===== HTTP Client =====
httpx==0.27.0
aiohttp==3.9.5

# ===== Stock Data =====
yfinance==0.2.38
pandas==2.2.2
numpy==1.26.4

# ===== WebSockets =====
websockets==12.0

# ===== Task Queue (background jobs) =====
celery==5.4.0
flower==2.0.1

# ===== AI / ML =====
openai==1.30.1
scikit-learn==1.4.2
ta==0.11.0          # Technical Analysis library

# ===== Utilities =====
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.3.0
orjson==3.10.3
